+++ srcjs/index.js
@@ -1,4 +1,5 @@
-import './exts/hostess.js';
+import './exts/attendant/attendant';
+import './exts/hostess/hostess.js';
 import waitress from './exts/waitress/waitress.js';
 import waiter from './exts/waiter/waiter.js';